Feat: Algebra methods
Added getDeterminant & getRank methods.
Changed main menu: now matrix is being created with size 3 and fill 0 instead of custom size initialization. Seems to be better.

There is the toUpperTriangularForm() method which is the key tool for the determinant and rank calculations. This tool may occur errors with hard numbers. Precision now is 3 digits after the dot.